+++ src/Logic/MessageFactory/StatementMessageFactory.php
@@ -6,6 +6,7 @@ use DemosEurope\DemosplanAddon\Utilities\AddonPath;
 use DemosEurope\DemosplanAddon\XBeteiligung\Configuration\Permissions\Features;
 use DemosEurope\DemosplanAddon\XBeteiligung\Exeption\NamespaceAdditionException;
 use DemosEurope\DemosplanAddon\XBeteiligung\Exeption\ProjectPrefixNotFoundException;
+use DemosEurope\DemosplanAddon\XBeteiligung\Logic\GeoreferenzierungConverter;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\SerializerFactory;
 use DemosEurope\DemosplanAddon\XBeteiligung\Soap\Schema\Kernmodul\AllgemeinerNameType;
 use DemosEurope\DemosplanAddon\XBeteiligung\Soap\Schema\Kernmodul\NameNatuerlichePersonType;
@@ -132,6 +133,16 @@ class StatementMessageFactory extends XBeteiligungResponseMessageFactory
         }
         // set Schlagwort
         $statement->setSchlagwort($statementCreated->getTags());
+        // set Einzeichnung
+        // TODO: May need to reproject coordinates to another coordinate system
+        // depending on XBeteiligung standard requirements (e.g., EPSG:25832, EPSG:4326)
+        $polygon = $statementCreated->getPolygon();
+        if ('' !== $polygon) {
+            $georeferenzierung = $this->georeferenzierungConverter->convertGeoJsonToGeoreferenzierung($polygon);
+            if ($georeferenzierung !== null) {
+                $statement->setGeoreferenzierung([$georeferenzierung]);
+            }
+        }
         $nachricht = new NachrichteninhaltAnonymousPHPType();
         $nachricht->setStellungnahme($statement);
         $nachricht->setVorgangsID($this->commonHelpers->uuid());
@@ -281,7 +292,6 @@ class StatementMessageFactory extends XBeteiligungResponseMessageFactory
         $simpleXML = simplexml_load_string($xml);
 
         $simpleXML->addAttribute('xmlns:xmlns:xsi', 'https://www.w3.org/2001/XMLSchema-instance');
-        $simpleXML->addAttribute('xmlns:xmlns:gml', 'https://www.opengis.net/gml/3.2');
 
         $result = $simpleXML->asXML();
         if (!is_string($result)) {